requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-PyMongo==2.3.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.0
MarkupSafe==1.1.1
matplotlib==3.3.2
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.51.0
traitlets==5.0.5
twython==3.8.2
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
wordcloud==1.8.0
